Consolidate some common code in TestCase types
The BrowserTestCase and PluginTestCase classes are intended to be available to plugins. This change uses traits to share common code between them.
