+++ tests/PluginTestCase.php
@@ -1,21 +1,17 @@
 <?php namespace October\Core\Tests;
 
-use System\Classes\UpdateManager;
-use System\Classes\PluginManager;
 use October\Rain\Database\Model as ActiveRecord;
 use October\Core\Tests\Concerns\CreatesApplication;
 use October\Core\Tests\Concerns\InteractsWithAuthentication;
+use October\Core\Tests\Concerns\RunsMigrations;
+use October\Core\Tests\Concerns\TestsPlugins;
 
 abstract class PluginTestCase extends TestCase
 {
     use CreatesApplication;
     use InteractsWithAuthentication;
-
-    /**
-     * @var array Cache for storing which plugins have been loaded
-     * and refreshed.
-     */
-    protected $pluginTestCaseLoadedPlugins = [];
+    use RunsMigrations;
+    use TestsPlugins;
 
     /**
      * Perform test case set up.
@@ -23,35 +19,20 @@ abstract class PluginTestCase extends TestCase
      */
     public function setUp() : void
     {
-        /*
-         * Force reload of October singletons
-         */
-        PluginManager::forgetInstance();
-        UpdateManager::forgetInstance();
+        $this->resetManagers();
 
-        /*
-         * Create application instance
-         */
+        // Create application
         parent::setUp();
 
-        /*
-         * Ensure system is up to date
-         */
-        $this->runOctoberUpCommand();
-
-        /*
-         * Detect plugin from test and autoload it
-         */
-        $this->pluginTestCaseLoadedPlugins = [];
-        $pluginCode = $this->guessPluginCodeFromTest();
-
-        if ($pluginCode !== false) {
-            $this->runPluginRefreshCommand($pluginCode, false);
+        // Ensure system is up to date
+        if ($this->usingTestDatabase) {
+            $this->runOctoberUpCommand();
         }
 
-        /*
-         * Disable mailer
-         */
+        // Detect a plugin and autoload it, if necessary
+        $this->detectPlugin();
+
+        // Disable mailer
         \Mail::pretend();
     }
 
@@ -62,88 +43,8 @@ abstract class PluginTestCase extends TestCase
     public function tearDown() : void
     {
         $this->flushModelEventListeners();
+
         parent::tearDown();
-        unset($this->app);
-    }
-
-    /**
-     * Migrate database using october:up command.
-     * @return void
-     */
-    protected function runOctoberUpCommand()
-    {
-        \Artisan::call('october:up');
-    }
-
-    /**
-     * Since the test environment has loaded all the test plugins
-     * natively, this method will ensure the desired plugin is
-     * loaded in the system before proceeding to migrate it.
-     * @return void
-     */
-    protected function runPluginRefreshCommand($code, $throwException = true)
-    {
-        if (!preg_match('/^[\w+]*\.[\w+]*$/', $code)) {
-            if (!$throwException) {
-                return;
-            }
-            throw new \Exception(sprintf('Invalid plugin code: "%s"', $code));
-        }
-
-        $manager = PluginManager::instance();
-        $plugin = $manager->findByIdentifier($code);
-
-        /*
-         * First time seeing this plugin, load it up
-         */
-        if (!$plugin) {
-            $namespace = '\\'.str_replace('.', '\\', strtolower($code));
-            $path = array_get($manager->getPluginNamespaces(), $namespace);
-
-            if (!$path) {
-                if (!$throwException) {
-                    return;
-                }
-                throw new \Exception(sprintf('Unable to find plugin with code: "%s"', $code));
-            }
-
-            $plugin = $manager->loadPlugin($namespace, $path);
-        }
-
-        /*
-         * Spin over dependencies and refresh them too
-         */
-        $this->pluginTestCaseLoadedPlugins[$code] = $plugin;
-
-        if (!empty($plugin->require)) {
-            foreach ((array) $plugin->require as $dependency) {
-                if (isset($this->pluginTestCaseLoadedPlugins[$dependency])) {
-                    continue;
-                }
-
-                $this->runPluginRefreshCommand($dependency);
-            }
-        }
-
-        /*
-         * Execute the command
-         */
-        \Artisan::call('plugin:refresh', ['name' => $code]);
-    }
-
-    /**
-     * Returns a plugin object from its code, useful for registering events, etc.
-     * @return PluginBase
-     */
-    protected function getPluginObject($code = null)
-    {
-        if ($code === null) {
-            $code = $this->guessPluginCodeFromTest();
-        }
-
-        if (isset($this->pluginTestCaseLoadedPlugins[$code])) {
-            return $this->pluginTestCaseLoadedPlugins[$code];
-        }
     }
 
     /**
@@ -172,25 +73,5 @@ abstract class PluginTestCase extends TestCase
         }
 
         ActiveRecord::flushEventListeners();
-    }
-
-    /**
-     * Locates the plugin code based on the test file location.
-     * @return string|bool
-     */
-    protected function guessPluginCodeFromTest()
-    {
-        $reflect = new \ReflectionClass($this);
-        $path = $reflect->getFilename();
-        $basePath = $this->app->pluginsPath();
-
-        $result = false;
-
-        if (strpos($path, $basePath) === 0) {
-            $result = ltrim(str_replace('\\', '/', substr($path, strlen($basePath))), '/');
-            $result = implode('.', array_slice(explode('/', $result), 0, 2));
-        }
-
-        return $result;
     }
 }
