+++ tests/concerns/RunsMigrations.php
@@ -1,16 +1,14 @@
 <?php namespace October\Core\Tests\Concerns;
 
-use Artisan;
-
 trait RunsMigrations
 {
     protected function runOctoberUpCommand()
     {
-        Artisan::call('october:up');
+        \Artisan::call('october:up');
     }
 
     protected function runOctoberDownCommand()
     {
-        Artisan::call('october:down --force');
+        \Artisan::call('october:down --force');
     }
 }
